fix: keep flutter previews independent of native runtime
