+++ docs/ui-styleguide.md
@@ -141,4 +141,5 @@ Every component ships a co-located `*.preview.dart` file so it can be developed 
 
 - Annotate preview functions with `@AleraPreview` (from `lib/src/design_system/alera_preview.dart`), which applies the Alera dark theme and ambient scaffolding automatically. Do not use the bare `@Preview`.
 - A preview function must be public and return a `Widget` or `WidgetBuilder`; its arguments and callbacks must be `const`/static.
-- Launch the previewer from the repo root with `flutter widget-preview start` (requires Chrome). Previews render on Flutter Web, so they cover UI only - terminal runtime, updater, and other native code paths are not previewable and stay out of the design system.
+- Launch the previewer with `flutter widget-preview start <absolute-project-path>` (the repository root for desktop, or its `mobile` directory). Passing an absolute path also keeps Flutter 3.47's scaffold restart anchored to the application when a pubspec changes. Previews render on Flutter Web, so they cover UI only - terminal runtime, updater, and other native code paths are not previewable and stay out of the design system. Quota previews use the presentational content; the application wrapper supplies its runtime actions.
+- The checked-in `.widget_preview/pubspec_overrides.yaml` files preserve the app's terminal forks in generated preview scaffolds. After `flutter widget-preview clean`, restore that file with `git restore -- .widget_preview/pubspec_overrides.yaml` from the corresponding application directory before starting again.
