+++ lib/src/features/agent_quota/presentation/agent_quota_hover_card.dart
@@ -1,4 +1,4 @@
-part of 'agent_quota_status_bar.dart';
+part of 'agent_quota_status_bar_content.dart';
 
 const double _quotaHoverCardWidth = 360;
 const double _quotaHoverCardMaxHeight = 480;
@@ -8,11 +8,13 @@ class _AgentQuotaHoverCard extends StatelessWidget {
     required this.snapshots,
     this.profileLabels = const <String, String>{},
     this.hostId = 'local',
+    this.actions = const AgentQuotaInlineActions(),
   });
 
   final List<AgentQuotaSnapshot> snapshots;
   final Map<String, String> profileLabels;
   final String hostId;
+  final AgentQuotaInlineActions actions;
 
   @override
   Widget build(BuildContext context) {
@@ -55,6 +57,7 @@ class _AgentQuotaHoverCard extends StatelessWidget {
                   snapshot: snapshot,
                   profileLabel: profileLabels[snapshot.key],
                   hostId: hostId,
+                  actions: actions,
                 ),
               ],
             ],
@@ -70,11 +73,13 @@ class _AgentQuotaHoverSection extends StatelessWidget {
     required this.snapshot,
     required this.profileLabel,
     required this.hostId,
+    this.actions = const AgentQuotaInlineActions(),
   });
 
   final AgentQuotaSnapshot snapshot;
   final String? profileLabel;
   final String hostId;
+  final AgentQuotaInlineActions actions;
 
   @override
   Widget build(BuildContext context) {
@@ -133,10 +138,10 @@ class _AgentQuotaHoverSection extends StatelessWidget {
             ],
           if (snapshot.provider == AgentQuotaProviderId.codex &&
               snapshot.rateLimitResetCredits != null)
-            _CodexResetCreditsPanel(hostId: hostId, snapshot: snapshot),
+            actions.buildCodexReset(hostId: hostId, snapshot: snapshot),
           if (_shouldOfferClaudeTui(snapshot)) ...<Widget>[
             const SizedBox(height: AleraTokens.space12),
-            _ClaudeTryWithTuiButton(hostId: hostId, snapshot: snapshot),
+            actions.buildClaudeTui(hostId: hostId, snapshot: snapshot),
           ],
         ],
       ),
@@ -147,77 +152,6 @@ class _AgentQuotaHoverSection extends StatelessWidget {
 bool _shouldOfferClaudeTui(AgentQuotaSnapshot snapshot) {
   return snapshot.provider == AgentQuotaProviderId.claude &&
       snapshot.status != AgentQuotaStatus.ok;
-}
-
-class _ClaudeTryWithTuiButton extends ConsumerStatefulWidget {
-  const _ClaudeTryWithTuiButton({required this.hostId, required this.snapshot});
-
-  final String hostId;
-  final AgentQuotaSnapshot snapshot;
-
-  @override
-  ConsumerState<_ClaudeTryWithTuiButton> createState() =>
-      _ClaudeTryWithTuiButtonState();
-}
-
-class _ClaudeTryWithTuiButtonState
-    extends ConsumerState<_ClaudeTryWithTuiButton> {
-  var _loading = false;
-
-  Future<void> _run() async {
-    if (_loading) {
-      return;
-    }
-    setState(() => _loading = true);
-    // The hover card can be disposed while the TUI request is in flight. Keep
-    // its longer-lived scope so the cached result still triggers a refresh.
-    final container = ProviderScope.containerOf(context, listen: false);
-    try {
-      final hostId = widget.hostId;
-      final targets = ref.read(sshTargetsProvider).value ?? const <SshTarget>[];
-      final target = hostId == 'local'
-          ? null
-          : targets.where((candidate) => candidate.id == hostId).firstOrNull;
-      await ref
-          .read(agentQuotaServiceProvider)
-          .fetchClaudeTui(
-            hostId: hostId,
-            target: target,
-            accountId: widget.snapshot.accountId,
-            displayName: widget.snapshot.displayName,
-          );
-    } on Object {
-      // Status bar refresh / next poll shows the updated error snapshot.
-    } finally {
-      container.invalidate(agentQuotaStateProvider);
-      if (mounted) {
-        setState(() => _loading = false);
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.centerRight,
-      child: TextButton(
-        onPressed: _loading ? null : _run,
-        style: TextButton.styleFrom(
-          foregroundColor: AleraTokens.accent,
-          padding: EdgeInsets.zero,
-          minimumSize: Size.zero,
-          tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-        ),
-        child: Text(
-          _loading ? 'Trying with TUI...' : 'Try With TUI',
-          style: Theme.of(context).textTheme.labelMedium?.copyWith(
-            color: _loading ? AleraTokens.foregroundFaint : AleraTokens.accent,
-            fontWeight: FontWeight.w600,
-          ),
-        ),
-      ),
-    );
-  }
 }
 
 class _QuotaHoverReading extends StatelessWidget {
