+++ lib/src/features/agent_quota/presentation/agent_quota_status_bar_readings.dart
@@ -1,4 +1,4 @@
-part of 'agent_quota_status_bar.dart';
+part of 'agent_quota_status_bar_content.dart';
 
 class _QuotaReading {
   const _QuotaReading({
