+++ mobile/lib/src/app/theme/alera_tokens.dart
@@ -28,6 +28,7 @@ abstract final class AleraTokens {
   static const double emptyStateMaxWidth = 520.0;
   static const double conversationMaxWidth = 760.0;
   static const double chatBubbleMaxWidth = 620.0;
+  static const Size previewPhoneSize = Size(390, 844);
 
   // Control radii match desktop so ported DS widgets look identical.
   static const double radiusSm = 4.0;
